Move revaultd binary src to src/bin/revauld/
We keep the configuration as a common lib as we'll reuse it for the CLI
