+++ Cargo.toml
@@ -4,6 +4,10 @@ version = "0.0.1"
 authors = ["Antoine Poinsot <darosior@protonmail.com>"]
 edition = "2018"
 
+[lib]
+name = "common"
+path = "src/common/lib.rs"
+
 [dependencies]
 revault_tx = { git = "https://github.com/darosior/revault", branch = "update_latest_miniscript" }
